requirements-gpu.txt
git+git://github.com/gamecraftCZ/pgdrive.git
tensorflow-gpu==2.4.1
keras-ncp==2.0.1
seaborn==0.11.1
numpy==1.19.3
matplotlib==3.3.4
h5py==2.10.0
opencv-python==4.5.1.48
keyboard==0.13.5
pydot==1.4.2
Flask==1.1.2